+++ crates/core/src/events.rs
@@ -23,10 +23,6 @@ impl_event!(InstalledEvent, {
     pub version: VersionSpec,
 });
 
-impl_event!(InstalledGlobalEvent, {
-    pub dependency: String,
-});
-
 impl_event!(UninstallingEvent, {
     pub version: VersionSpec,
 });