+++ crates/core/src/tool.rs
@@ -55,7 +55,6 @@ pub struct Tool {
     pub on_created_shims: Emitter<CreatedShimsEvent>,
     pub on_installing: Emitter<InstallingEvent>,
     pub on_installed: Emitter<InstalledEvent>,
-    pub on_installed_global: Emitter<InstalledGlobalEvent>,
     pub on_resolved_version: Emitter<ResolvedVersionEvent>,
     pub on_uninstalling: Emitter<UninstallingEvent>,
     pub on_uninstalled: Emitter<UninstalledEvent>,
@@ -96,7 +95,6 @@ impl Tool {
             on_created_shims: Emitter::new(),
             on_installing: Emitter::new(),
             on_installed: Emitter::new(),
-            on_installed_global: Emitter::new(),
             on_resolved_version: Emitter::new(),
             on_uninstalling: Emitter::new(),
             on_uninstalled: Emitter::new(),
@@ -977,44 +975,6 @@ impl Tool {
         );
 
         Ok(true)
-    }
-
-    /// Install a global dependency/package for the tool.
-    pub async fn install_global(&self, dependency: &str) -> miette::Result<bool> {
-        let globals_dir = self.get_globals_bin_dir();
-
-        if !self.plugin.has_func("install_global") || globals_dir.is_none() {
-            return Ok(false);
-        }
-
-        if is_offline() {
-            return Err(ProtoError::InternetConnectionRequired.into());
-        }
-
-        let result: InstallGlobalOutput = self.plugin.call_func_with(
-            "install_global",
-            InstallGlobalInput {
-                dependency: dependency.to_owned(),
-                globals_dir: self.to_virtual_path(globals_dir.as_ref().unwrap()),
-                context: self.create_context(),
-            },
-        )?;
-
-        if !result.installed {
-            return Err(ProtoError::InstallFailed {
-                tool: dependency.to_owned(),
-                error: result.error.unwrap_or_default(),
-            }
-            .into());
-        }
-
-        self.on_installed_global
-            .emit(InstalledGlobalEvent {
-                dependency: dependency.to_owned(),
-            })
-            .await?;
-
-        Ok(result.installed)
     }
 
     /// Uninstall the tool by deleting the current install directory.